Added accelerator flag checking functionality and automatic accelerator selection:
Following the nested dictionary format of the Accelerator Registry, the
handler checks with pytorch backends and the availability of the
accelerator within the native TA registry.

If both checks succeed, the program makes a call to the Accelerator
Registry module for the accelerator class, adaped from the
BaseAccelerator base class.

If either check fails, the program defaults to "cpu" usage.